Austin updates

+++ rsync.sh
@@ -40,18 +40,31 @@ deployweb(){
         DRY="n"
 		echo "DRY DEPLOY"
     fi
-	rsync -avz${DRY} --delete --exclude '__pycache__/' --exclude 'migrations/' website/ porchfest:/home/django/porchfest/website/
-	rsync -avz${DRY} static/ porchfest:/home/django/porchfest/static/
+	if [ "$1" = "rebuild" ]; then
+		rsync -avz${DRY} --delete --exclude '__pycache__/' --exclude 'migrations/' website/ porchfest:/home/django/porchfest/website/
+		rsync -avz${DRY} static/ porchfest:/home/django/porchfest/static/
 
-	if [ "$1" != "--dry" ]; then
-		ssh porchfest << "EOF"
-		cd /home/django/porchfest
-		source venv/bin/activate
-		python manage.py collectstatic --noinput
-		systemctl reload nginx
-		systemctl restart gunicorn
+		if [ "$1" != "--dry" ]; then
+			ssh porchfest << "EOF"
+			cd /home/django/porchfest
+			source venv/bin/activate
+			python manage.py collectstatic --noinput
+			systemctl reload nginx
+			systemctl restart gunicorn
 EOF
+		fi
+	else
+		rsync -avz${DRY} --delete --exclude '__pycache__/' --exclude 'migrations/' website/ porchfest:/home/django/porchfest/website/
 	fi
+}
+
+deploypyd(){
+	DRY=""
+    if [ "$1" = "--dry" ]; then
+        DRY="n"
+		echo "DRY DEPLOY"
+    fi
+	rsync -avz${DRY} --delete --exclude '__pycache__/' --exclude 'migrations/' planyourday/ porchfest:/home/django/porchfest/planyourday/
 }
 
 deploydash(){
@@ -90,6 +103,9 @@ elif [ "$1" == "deploystatic" ]; then
 elif [ "$1" == "deploydash" ]; then
 	shift
 	deploydash "$@"
+elif [ "$1" == "deploypyd" ]; then
+	shift
+	deploypyd "$@"
 elif [ "$1" == "deploycore" ]; then
 	shift
 	deploycore "$@"
